// cotown/src/app/time-chart/models/time-chart-line.model.ts
export class TimeChartLine {
    // Set
    code: string = '';
    color: string = '';
    lock: boolean = false;
    datefrom: Date = new Date();
    dateto: Date = new Date();
    type: string = '';
    text: string = '';
    tooltip: string = '';

    // Calculated
    from: number = 0;
    to: number = 0;
    styles: string = '';
}